+++ packages/elf-service/tests/acceptance/docs_extension_v1.rs
@@ -1,5 +1,9 @@
+mod excerpts;
+mod indexing;
+mod l0_search;
 mod lifecycle;
 mod search_filters;
+mod validation_rejections;
 
 use std::{
 	collections::HashSet,
@@ -29,9 +33,8 @@ use uuid::Uuid;
 use crate::acceptance::{self, SpyExtractor, StubEmbedding, StubRerank, chunking::ChunkingConfig};
 use elf_config::EmbeddingProviderConfig;
 use elf_service::{
-	AddNoteInput, AddNoteRequest, DocsExcerptsGetRequest, DocsGetRequest, DocsPutRequest,
-	DocsPutResponse, DocsSearchL0Request, ElfService, Error, PayloadLevel, Providers,
-	SearchRequest, TextQuoteSelector, docs::DocRetrievalTrajectory,
+	DocsExcerptsGetRequest, DocsGetRequest, DocsPutRequest, DocsPutResponse, DocsSearchL0Request,
+	ElfService, Providers, TextQuoteSelector, docs::DocRetrievalTrajectory,
 };
 use elf_storage::{db::Db, qdrant::QdrantStore};
 use elf_testkit::TestDatabase;
@@ -317,276 +320,6 @@ async fn cleanup_docs_filter_fixture(
 	test_db.cleanup().await.expect("Failed to cleanup test database.");
 }
 
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL (or ELF_QDRANT_GRPC_URL) to run."]
-async fn docs_put_rejects_non_english_source_ref() {
-	let Some(test_db) = acceptance::test_db().await else {
-		eprintln!("Skipping docs_extension_v1; set ELF_PG_DSN to run this test.");
-
-		return;
-	};
-	let Some(qdrant_url) = acceptance::test_qdrant_url() else {
-		eprintln!(
-			"Skipping docs_extension_v1; set ELF_QDRANT_URL (or ELF_QDRANT_GRPC_URL) to run this test."
-		);
-
-		return;
-	};
-	let collection = test_db.collection_name("elf_acceptance");
-	let docs_collection = test_db.collection_name("elf_acceptance_docs");
-	let cfg = acceptance::test_config(
-		test_db.dsn().to_string(),
-		qdrant_url,
-		4_096,
-		collection,
-		docs_collection,
-	);
-	let providers = Providers::new(
-		Arc::new(StubEmbedding { vector_dim: 4_096 }),
-		Arc::new(StubRerank),
-		Arc::new(SpyExtractor {
-			calls: Arc::new(Default::default()),
-			payload: serde_json::json!({ "notes": [] }),
-		}),
-	);
-	let service =
-		acceptance::build_service(cfg, providers).await.expect("Failed to build service.");
-	let result = service
-		.docs_put(DocsPutRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "owner".to_string(),
-			scope: "project_shared".to_string(),
-			doc_type: None,
-			title: Some("Docs rejection sample".to_string()),
-			write_policy: None,
-			source_ref: serde_json::json!({
-				"schema": "doc_source_ref/v1",
-				"doc_type": "knowledge",
-				"ts": "2026-02-25T12:00:00Z",
-				"notes": "你好"
-			}),
-			content: TEST_CONTENT.to_string(),
-		})
-		.await;
-
-	match result {
-		Err(Error::NonEnglishInput { field }) => {
-			assert_eq!(field, "$.source_ref[\"notes\"]");
-		},
-		other => panic!("Expected NonEnglishInput, got {other:?}"),
-	}
-
-	test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL (or ELF_QDRANT_GRPC_URL) to run."]
-async fn docs_put_rejects_missing_and_invalid_source_ref() {
-	let Some(test_db) = acceptance::test_db().await else {
-		eprintln!("Skipping docs_extension_v1; set ELF_PG_DSN to run this test.");
-
-		return;
-	};
-	let Some(qdrant_url) = acceptance::test_qdrant_url() else {
-		eprintln!(
-			"Skipping docs_extension_v1; set ELF_QDRANT_URL (or ELF_QDRANT_GRPC_URL) to run this test."
-		);
-
-		return;
-	};
-	let collection = test_db.collection_name("elf_acceptance");
-	let docs_collection = test_db.collection_name("elf_acceptance_docs");
-	let cfg = acceptance::test_config(
-		test_db.dsn().to_string(),
-		qdrant_url,
-		4_096,
-		collection,
-		docs_collection,
-	);
-	let providers = Providers::new(
-		Arc::new(StubEmbedding { vector_dim: 4_096 }),
-		Arc::new(StubRerank),
-		Arc::new(SpyExtractor {
-			calls: Arc::new(Default::default()),
-			payload: serde_json::json!({ "notes": [] }),
-		}),
-	);
-	let service =
-		acceptance::build_service(cfg, providers).await.expect("Failed to build service.");
-	let result = service
-		.docs_put(DocsPutRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "owner".to_string(),
-			scope: "project_shared".to_string(),
-			doc_type: None,
-			title: Some("Docs rejection sample".to_string()),
-			write_policy: None,
-			source_ref: serde_json::json!("legacy-shape"),
-			content: TEST_CONTENT.to_string(),
-		})
-		.await;
-
-	match result {
-		Err(Error::InvalidRequest { message }) => {
-			assert!(message.contains("source_ref must be a JSON object"));
-		},
-		other => panic!("Expected InvalidRequest for non-object source_ref, got {other:?}"),
-	}
-
-	let result = service
-		.docs_put(DocsPutRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "owner".to_string(),
-			scope: "project_shared".to_string(),
-			doc_type: None,
-			title: Some("Docs rejection sample".to_string()),
-			write_policy: None,
-			source_ref: serde_json::json!({
-				"schema": "source_ref/v1",
-				"doc_type": "knowledge",
-				"ts": "2026-02-25T12:00:00Z",
-			}),
-			content: TEST_CONTENT.to_string(),
-		})
-		.await;
-
-	match result {
-		Err(Error::InvalidRequest { message }) => {
-			assert!(message.contains("doc_source_ref/v1"));
-		},
-		other => panic!("Expected InvalidRequest for wrong source_ref schema, got {other:?}"),
-	}
-
-	test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL (or ELF_QDRANT_GRPC_URL) to run."]
-async fn docs_search_l0_requires_qdrant_payload_indexes_for_filters() {
-	let Some(ctx) = setup_docs_context().await else { return };
-	let DocsContext { test_db, service } = ctx;
-	let doc = put_test_doc(&service).await;
-	let (handle, shutdown) = spawn_doc_worker(&service).await;
-
-	assert!(
-		wait_for_doc_outbox_done(&service.db.pool, doc.doc_id, std::time::Duration::from_secs(15))
-			.await,
-		"Expected doc outbox to reach DONE."
-	);
-
-	verify_docs_qdrant_filter_indexes(&service).await;
-
-	let _ = shutdown.send(());
-
-	handle.abort();
-
-	let _ = handle.await;
-
-	test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL (or ELF_QDRANT_GRPC_URL) to run."]
-async fn docs_search_l0_projects_source_ref_payload_fields() {
-	let Some(ctx) = setup_docs_context().await else { return };
-	let DocsContext { test_db, service } = ctx;
-	let source_ts = "2025-01-01T10:00:00Z";
-	let cases = [
-		(
-			"chat",
-			"Docs chat source ref sample",
-			serde_json::json!({
-				"schema": "doc_source_ref/v1",
-				"doc_type": "chat",
-				"ts": source_ts,
-				"thread_id": "thread-42",
-				"role": "assistant"
-			}),
-			("thread_id", "thread-42"),
-			["domain", "repo"],
-		),
-		(
-			"search",
-			"Docs search source ref sample",
-			serde_json::json!({
-				"schema": "doc_source_ref/v1",
-				"doc_type": "search",
-				"ts": source_ts,
-				"query": "What is payload indexing?",
-				"url": "https://docs.example.com/search",
-				"domain": "docs.example.com",
-				"provider": "web"
-			}),
-			("domain", "docs.example.com"),
-			["thread_id", "repo"],
-		),
-		(
-			"dev",
-			"Docs dev source ref sample",
-			serde_json::json!({
-				"schema": "doc_source_ref/v1",
-				"doc_type": "dev",
-				"ts": source_ts,
-				"repo": "elf-org/docs",
-				"commit_sha": "9f0a3f4c4eb58bfcf4a5f4f9d0c7be0e13c2f8d19"
-			}),
-			("repo", "elf-org/docs"),
-			["thread_id", "domain"],
-		),
-	];
-	let mut docs = Vec::new();
-
-	for (doc_type, title, source_ref, expected_present, expected_absent) in cases {
-		let doc = put_test_doc_with(
-			&service,
-			"owner",
-			"project_shared",
-			Some(doc_type),
-			title,
-			source_ref,
-			TEST_CONTENT,
-		)
-		.await;
-
-		docs.push((doc.doc_id, expected_present, expected_absent));
-	}
-
-	let (handle, shutdown) = spawn_doc_worker(&service).await;
-
-	for (doc_id, expected_present, expected_absent) in &docs {
-		assert!(
-			wait_for_doc_outbox_done(&service.db.pool, *doc_id, std::time::Duration::from_secs(15))
-				.await,
-			"Expected doc outbox to reach DONE."
-		);
-
-		let point = fetch_first_doc_chunk_point(&service, *doc_id)
-			.await
-			.expect("Expected doc chunk point in Qdrant.");
-
-		assert_eq!(point.payload.get("doc_ts").and_then(payload_string), Some(source_ts));
-		assert_eq!(
-			point.payload.get(expected_present.0).and_then(payload_string),
-			Some(expected_present.1)
-		);
-
-		for key in expected_absent {
-			assert!(!point.payload.contains_key(*key));
-		}
-	}
-
-	_ = shutdown.send(());
-
-	handle.abort();
-
-	let _ = handle.await;
-
-	test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
 async fn setup_docs_context() -> Option<DocsContext> {
 	let Some(test_db) = acceptance::test_db().await else {
 		eprintln!("Skipping docs_extension_v1; set ELF_PG_DSN to run this test.");
@@ -683,298 +416,6 @@ async fn put_test_doc(service: &ElfService) -> DocsPutResponse {
 		TEST_CONTENT,
 	)
 	.await
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL (or ELF_QDRANT_GRPC_URL) to run."]
-async fn docs_search_l0_returns_pointer_and_explain_trajectory() {
-	let Some(ctx) = setup_docs_context().await else { return };
-	let DocsContext { test_db, service } = ctx;
-	let doc = put_test_doc(&service).await;
-	let (handle, shutdown) = spawn_doc_worker(&service).await;
-
-	assert!(
-		wait_for_doc_outbox_done(&service.db.pool, doc.doc_id, std::time::Duration::from_secs(15))
-			.await,
-		"Expected doc outbox to reach DONE."
-	);
-
-	let results = service
-		.docs_search_l0(DocsSearchL0Request {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			caller_agent_id: "reader".to_string(),
-			scope: None,
-			status: None,
-			doc_type: None,
-			sparse_mode: None,
-			domain: None,
-			repo: None,
-			agent_id: None,
-			thread_id: None,
-			updated_after: None,
-			updated_before: None,
-			ts_gte: None,
-			ts_lte: None,
-			read_profile: "private_plus_project".to_string(),
-			query: "peregrine".to_string(),
-			top_k: Some(5),
-			candidate_k: Some(20),
-			explain: Some(true),
-		})
-		.await
-		.expect("Failed to search docs.");
-
-	assert_eq!(
-		results.trajectory.as_ref().map(|trajectory| trajectory.schema.as_str()),
-		Some("doc_retrieval_trajectory/v1")
-	);
-	assert!(results.trajectory.is_some());
-	assert!(!results.items.is_empty());
-	assert!(results.items[0].pointer.schema == "source_ref/v1");
-	assert!(!results.items[0].pointer.reference.doc_id.is_nil());
-	assert!(!results.items[0].pointer.reference.chunk_id.is_nil());
-	assert_eq!(
-		results.items[0].pointer.reference.source_record_id,
-		results.items[0].pointer.reference.doc_id
-	);
-	assert_eq!(
-		results.items[0].pointer.reference.source_span_id,
-		results.items[0].pointer.locator.span_id
-	);
-	assert_eq!(results.items[0].pointer.resolver, "elf_doc_ext/v1");
-	assert!(!results.items[0].pointer.locator.span_id.is_nil());
-	assert!(!results.trace_id.is_nil());
-
-	let _ = shutdown.send(());
-
-	handle.abort();
-
-	let _ = handle.await;
-
-	drop(service);
-
-	test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL (or ELF_QDRANT_GRPC_URL) to run."]
-async fn docs_search_l0_note_pointer_roundtrip_hydrates_doc() {
-	let Some(ctx) = setup_docs_context().await else { return };
-	let DocsContext { test_db, service } = ctx;
-	let doc = put_test_doc(&service).await;
-	let (handle, shutdown) = spawn_doc_worker(&service).await;
-
-	assert!(
-		wait_for_doc_outbox_done(&service.db.pool, doc.doc_id, std::time::Duration::from_secs(15))
-			.await,
-		"Expected doc outbox to reach DONE."
-	);
-
-	let (source_ref, source_ref_doc_id, source_ref_chunk_id) =
-		fetch_docs_pointer_source_ref(&service).await;
-	let note_id = add_note_with_pointer_source_ref(&service, source_ref.clone()).await;
-
-	assert!(
-		wait_for_note_outbox_done(&service.db.pool, note_id, std::time::Duration::from_secs(15))
-			.await,
-		"Expected note outbox to reach DONE."
-	);
-
-	let search_results = service
-		.search_raw_quick(SearchRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "agent".to_string(),
-			token_id: None,
-			read_profile: "private_only".to_string(),
-			payload_level: PayloadLevel::L2,
-			query: "peregrine".to_string(),
-			top_k: Some(5),
-			candidate_k: Some(20),
-			filter: None,
-			record_hits: Some(false),
-			ranking: None,
-		})
-		.await
-		.expect("Failed to search note with doc pointer source_ref.");
-	let has_pointer_source_ref =
-		search_results.items.into_iter().any(|item| item.source_ref == source_ref);
-
-	assert!(
-		has_pointer_source_ref,
-		"Expected search result to include note with pointer source_ref."
-	);
-
-	let excerpt = service
-		.docs_excerpts_get(DocsExcerptsGetRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "reader".to_string(),
-			read_profile: "private_plus_project".to_string(),
-			doc_id: source_ref_doc_id,
-			level: "L1".to_string(),
-			chunk_id: Some(source_ref_chunk_id),
-			quote: None,
-			position: None,
-			explain: None,
-		})
-		.await
-		.expect("Failed to hydrate excerpt from pointer source_ref.");
-
-	assert!(excerpt.verification.verified);
-
-	let _ = shutdown.send(());
-
-	handle.abort();
-
-	let _ = handle.await;
-
-	drop(service);
-
-	test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-async fn fetch_docs_pointer_source_ref(service: &ElfService) -> (serde_json::Value, Uuid, Uuid) {
-	let search = service
-		.docs_search_l0(DocsSearchL0Request {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			caller_agent_id: "reader".to_string(),
-			scope: None,
-			status: None,
-			doc_type: None,
-			sparse_mode: None,
-			domain: None,
-			repo: None,
-			agent_id: None,
-			thread_id: None,
-			updated_after: None,
-			updated_before: None,
-			ts_gte: None,
-			ts_lte: None,
-			read_profile: "private_plus_project".to_string(),
-			query: "peregrine".to_string(),
-			top_k: Some(5),
-			candidate_k: Some(20),
-			explain: None,
-		})
-		.await
-		.expect("Failed to search docs for source_ref pointer.");
-
-	assert!(!search.items.is_empty(), "Expected docs_search_l0 to return source_ref pointer.");
-
-	let pointer = search.items[0].pointer.clone();
-	let source_ref =
-		serde_json::to_value(&pointer).expect("Failed to serialize docs_search_l0 pointer.");
-
-	(source_ref, pointer.reference.doc_id, pointer.reference.chunk_id)
-}
-
-async fn add_note_with_pointer_source_ref(
-	service: &ElfService,
-	source_ref: serde_json::Value,
-) -> Uuid {
-	let note = service
-		.add_note(AddNoteRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "agent".to_string(),
-			scope: "agent_private".to_string(),
-			notes: vec![AddNoteInput {
-				r#type: "fact".to_string(),
-				key: Some("doc_pointer_note".to_string()),
-				text: "Peregrine note for source_ref hydration check.".to_string(),
-				structured: None,
-				importance: 0.5,
-				confidence: 0.9,
-				ttl_days: None,
-				source_ref,
-				write_policy: None,
-			}],
-		})
-		.await
-		.expect("Failed to add note from docs pointer.");
-
-	note.results[0].note_id.expect("Expected note_id in add_note result.")
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL (or ELF_QDRANT_GRPC_URL) to run."]
-async fn docs_excerpts_get_supports_l0_and_returns_locator_and_optional_trajectory() {
-	let Some(ctx) = setup_docs_context().await else { return };
-	let DocsContext { test_db, service } = ctx;
-	let doc = put_test_doc(&service).await;
-	let (handle, shutdown) = spawn_doc_worker(&service).await;
-
-	assert!(
-		wait_for_doc_outbox_done(&service.db.pool, doc.doc_id, std::time::Duration::from_secs(15))
-			.await,
-		"Expected doc outbox to reach DONE."
-	);
-
-	let excerpt = service
-		.docs_excerpts_get(DocsExcerptsGetRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "reader".to_string(),
-			read_profile: "private_plus_project".to_string(),
-			doc_id: doc.doc_id,
-			level: "L0".to_string(),
-			chunk_id: None,
-			quote: Some(TextQuoteSelector {
-				exact: "Keyword: peregrine.".to_string(),
-				prefix: Some("evidence. ".to_string()),
-				suffix: Some("\nSecond".to_string()),
-			}),
-			position: None,
-			explain: Some(true),
-		})
-		.await
-		.expect("Failed to hydrate excerpt.");
-
-	assert_eq!(excerpt.locator.selector_kind, "quote");
-	assert!(excerpt.locator.match_end_offset > excerpt.locator.match_start_offset);
-	assert!(!excerpt.locator.span_id.is_nil());
-	assert!(excerpt.excerpt.len() <= 256);
-	assert!(excerpt.trajectory.is_some());
-	assert_eq!(
-		excerpt.trajectory.as_ref().map(|trajectory| trajectory.schema.as_str()),
-		Some("doc_retrieval_trajectory/v1")
-	);
-	assert!(!excerpt.trace_id.is_nil());
-
-	let no_explain = service
-		.docs_excerpts_get(DocsExcerptsGetRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "reader".to_string(),
-			read_profile: "private_plus_project".to_string(),
-			doc_id: doc.doc_id,
-			level: "L0".to_string(),
-			chunk_id: None,
-			quote: Some(TextQuoteSelector {
-				exact: "Keyword: peregrine.".to_string(),
-				prefix: Some("evidence. ".to_string()),
-				suffix: Some("\nSecond".to_string()),
-			}),
-			position: None,
-			explain: Some(false),
-		})
-		.await
-		.expect("Failed to hydrate excerpt.");
-
-	assert!(no_explain.trajectory.is_none());
-
-	let _ = shutdown.send(());
-
-	handle.abort();
-
-	let _ = handle.await;
-
-	drop(service);
-
-	test_db.cleanup().await.expect("Failed to cleanup test database.");
 }
 
 async fn put_test_doc_with(
